transform: move `Cardinality` code from `attribute`
